downloader: Store the initial revision of a working tree
Until now the initial, probably symbolic, revision that was used to
create the working tree got lost when returning from download().
However, that revision is beneficial to known for provenance reasons.

Note that the initial revision should not be returned from getInfo() as
that function should always reflect the current state of the working
tree.

Signed-off-by: Sebastian Schuberth <sebastian.schuberth@bosch.io>