OrtResult: Add a labels property
Add a property for custom labels to the OrtResult. These labels will
later be made available to the evaluator rules and the reporters. ORT
itself will not use those lables directly, but they can be used by users
in evaluator rules or the notice reporters.

For example, if there are tiny differences in how certain licenses are
handled in the evaluator rules based on if the software is distributed
internally or externally, the users could set the label "dist=internal"
or "dist=external" and access this in the rules.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@here.com>